Plume 3D Visualization/requirements.txt
numpy>=1.23
pandas>=1.5
matplotlib>=3.5
PyYAML>=6.0
scipy>=1.10
rasterio>=1.3
noise>=1.2
vtk==9.3.0
mayavi==4.8.1
traits>=6.4
PyQt5>=5.15
